+++ app/actor_register/GlobalChatActorRegister.scala
@@ -1,13 +1,14 @@
-package actorRegister
+package actor_register
 
-import actorRegister.GlobalChatActorRegister.actorRegister
+import actor_register.GlobalChatActorRegister.actorRegister
 import akka.actor.ActorSystem
 
-object MatchListActorRegister {
+object GlobalChatActorRegister {
   var actorRegister: scala.collection.mutable.Map[Int, String] = scala.collection.mutable.Map[Int, String]()
 }
 
-class MatchListActorRegister(actorSystem: ActorSystem) {
+class GlobalChatActorRegister(actorSystem: ActorSystem) {
+
   def unregisterClosedSocket(actorPathForDelete: String): Unit = {
     actorRegister.foreach(actor =>
       if(actor._2 == actorPathForDelete) {
